refactor: Update pylint configuration to allow negative min-public-methods value
Update the `.pylintrc` configuration file to allow a negative value for the `min-public-methods` rule. This change is necessary to accommodate the refactoring of the codebase, which involves reducing the number of public methods in certain classes. By allowing a negative value, the rule will be disabled, allowing for more flexibility in defining the minimum number of public methods for a class.